+++ mupen64plus-core/src/r4300/r4300.h
@@ -30,12 +30,12 @@ extern precomp_instr *PC;
 extern precomp_block *blocks[0x100000], *actual;
 
 extern int stop, llbit, rompause;
-extern long long int reg[32], hi, lo;
-extern long long int local_rs;
+extern int64_t reg[32], hi, lo;
+extern int64_t local_rs;
 extern unsigned int reg_cop0[32];
 extern double *reg_cop1_double[32];
 extern float *reg_cop1_simple[32];
-extern long long int reg_cop1_fgr_64[32];
+extern int64_t reg_cop1_fgr_64[32];
 extern int FCR0, FCR31;
 extern tlb tlb_e[32];
 extern unsigned int delay_slot, skip_jump, dyna_interp, op;
